Migrate issuecomment table to i64 ids
2 years after the stopgap fix in f9d84fa85b54296, this writes a
one-way migration to move to bigint columns for all the issuecomment
IDs. Presuming we like this approach, it would be a good idea to apply
the same to other GitHub IDs (issues, users, etc.).

The migration has been tested on a data export from production postgres,
locally it takes ~7 seconds to run. I'm not sure if it's safe to run it
while rfcbot is actively running (in particular, what happens if the
diesel schema in the running rfcbot disagrees with the database). We may
want to try to stop the app, run the migration, and then start it again
(downtime should be comparatively short though not zero). rfcbot isn't
currently responding to any new comments I think so it's in practice
nearly down anyway.

+++ src/domain/rfcbot.rs
@@ -7,10 +7,8 @@ use super::schema::*;
 pub struct NewPoll<'a> {
     pub fk_issue: i32,
     pub fk_initiator: i32,
-    //#[serde(serialize_with = "super::unsigned")]
-    pub fk_initiating_comment: i32,
-    //#[serde(serialize_with = "super::unsigned")]
-    pub fk_bot_tracking_comment: i32,
+    pub fk_initiating_comment: i64,
+    pub fk_bot_tracking_comment: i64,
     pub poll_question: &'a str,
     pub poll_created_at: NaiveDateTime,
     pub poll_closed: bool,
@@ -25,10 +23,8 @@ pub struct Poll {
     pub id: i32,
     pub fk_issue: i32,
     pub fk_initiator: i32,
-    #[serde(serialize_with = "super::unsigned")]
-    pub fk_initiating_comment: i32,
-    #[serde(serialize_with = "super::unsigned")]
-    pub fk_bot_tracking_comment: i32,
+    pub fk_initiating_comment: i64,
+    pub fk_bot_tracking_comment: i64,
     pub poll_question: String,
     pub poll_created_at: NaiveDateTime,
     pub poll_closed: bool,
@@ -40,11 +36,9 @@ pub struct Poll {
 pub struct NewFcpProposal<'a> {
     pub fk_issue: i32,
     pub fk_initiator: i32,
-    //#[serde(serialize_with = "super::unsigned")]
-    pub fk_initiating_comment: i32,
+    pub fk_initiating_comment: i64,
     pub disposition: &'a str,
-    //#[serde(serialize_with = "super::unsigned")]
-    pub fk_bot_tracking_comment: i32,
+    pub fk_bot_tracking_comment: i64,
     pub fcp_start: Option<NaiveDateTime>,
     pub fcp_closed: bool,
 }
@@ -77,11 +71,9 @@ pub struct FcpProposal {
     pub id: i32,
     pub fk_issue: i32,
     pub fk_initiator: i32,
-    #[serde(serialize_with = "super::unsigned")]
-    pub fk_initiating_comment: i32,
+    pub fk_initiating_comment: i64,
     pub disposition: String,
-    #[serde(serialize_with = "super::unsigned")]
-    pub fk_bot_tracking_comment: i32,
+    pub fk_bot_tracking_comment: i64,
     pub fcp_start: Option<NaiveDateTime>,
     pub fcp_closed: bool,
 }
@@ -110,9 +102,9 @@ pub struct FcpReviewRequest {
 pub struct NewFcpConcern<'a> {
     pub fk_proposal: i32,
     pub fk_initiator: i32,
-    pub fk_resolved_comment: Option<i32>,
+    pub fk_resolved_comment: Option<i64>,
     pub name: &'a str,
-    pub fk_initiating_comment: i32,
+    pub fk_initiating_comment: i64,
 }
 
 #[derive(AsChangeset, Clone, Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Queryable)]
@@ -121,9 +113,9 @@ pub struct FcpConcern {
     pub id: i32,
     pub fk_proposal: i32,
     pub fk_initiator: i32,
-    pub fk_resolved_comment: Option<i32>,
+    pub fk_resolved_comment: Option<i64>,
     pub name: String,
-    pub fk_initiating_comment: i32,
+    pub fk_initiating_comment: i64,
 }
 
 #[derive(Clone, Debug, Eq, Insertable, Ord, PartialEq, PartialOrd)]
@@ -132,7 +124,7 @@ pub struct NewFeedbackRequest {
     pub fk_initiator: i32,
     pub fk_requested: i32,
     pub fk_issue: i32,
-    pub fk_feedback_comment: Option<i32>,
+    pub fk_feedback_comment: Option<i64>,
 }
 
 #[derive(AsChangeset, Clone, Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Queryable)]
@@ -142,5 +134,5 @@ pub struct FeedbackRequest {
     pub fk_initiator: i32,
     pub fk_requested: i32,
     pub fk_issue: i32,
-    pub fk_feedback_comment: Option<i32>,
+    pub fk_feedback_comment: Option<i64>,
 }
